+++ Frends.SFTP.DownloadFiles/Frends.SFTP.DownloadFiles/Definitions/FileTransporter.cs
@@ -28,7 +28,9 @@ internal class FileTransporter
         _result = new List<SingleFileTransferResult>();
         _filePaths = ConvertObjectToStringArray(context.Source.FilePaths);
 
-        SourceDirectoryWithMacrosExtended = _renamingPolicy.ExpandDirectoryForMacros(context.Source.Directory);
+        if (_filePaths == null || !_filePaths.Any())
+            SourceDirectoryWithMacrosExtended = _renamingPolicy.ExpandDirectoryForMacros(context.Source.Directory);
+
         DestinationDirectoryWithMacrosExtended = _renamingPolicy.ExpandDirectoryForMacros(context.Destination.Directory);
     }
 
@@ -396,9 +398,18 @@ internal class FileTransporter
 
         if (_filePaths != null)
         {
-            fileItems = _filePaths.Select(p => new FileItem(p) { Name = p }).ToList();
+            var items = _filePaths.Select(p => new FileItem(p) { Name = p }).ToList();
+            foreach (var file in items)
+            {
+                if (!client.Exists(file.FullPath))
+                    _logger.NotifyError(_batchContext, $"File does not exist: '{file.FullPath}", new FileNotFoundException());
+                else
+                    fileItems.Add(file);
+            }
+               
             if (fileItems.Any()) return new Tuple<List<FileItem>, bool>(fileItems, true);
-            return new Tuple<List<FileItem>, bool>(fileItems, false);
+
+            return new Tuple<List<FileItem>, bool>(fileItems, true);
         }
 
         // Return empty list and success.false value if source directory doesn't exists.